fix: xpath for individual link selection
HN added a class `submission` to `tr` tag for each link and the xpath selection
was matching on a `tr` with class `athing`. To avoid this problem in the future
this commit introduces a fix that selects all the `tr` elements that contain a
class `athing`.
