+++ HNWidget/WidgetConfigurationIntent.swift
@@ -35,25 +35,3 @@ enum UpdateEveryIntent: Int, AppEnum {
     }
 }
 
-extension HNPageFetcher.HNList: AppEnum {
-    static var typeDisplayRepresentation: TypeDisplayRepresentation {
-        TypeDisplayRepresentation(name: "List")
-    }
-    
-    typealias Display = [HNPageFetcher.HNList:DisplayRepresentation]
-    static var caseDisplayRepresentations: Display {
-        [
-            .active : "active",
-            .asknew : "asknew",
-            .best : "best",
-            .classic : "classic",
-            .home : "home",
-            .invited : "invited",
-            .launches: "launches",
-            .noobstories: "noobstories",
-            .pool: "pool",
-            .shownew: "shownew",
-            .whoishiring: "whoishiring"
-        ]
-    }
-}
